requirements.txt
torch
sentencepiece
datasets
deepspeed==0.9.1
# accelerate==0.18.0
git+https://github.com/huggingface/accelerate.git@e06e7b35e7c49563ed70f31b78126bf70b0252b9  # this version fixes a bug about gradient accumulation in 0.18.0 
git+https://github.com/huggingface/peft.git@49a20c16dcd9de5716feee717e8eb742efb9eff9
bitsandbytes==0.37.2
evaluate==0.4.0
tokenizers==0.13.3
protobuf==3.20.0
transformers>=4.28.1
openai
tiktoken
rouge_score
wandb
gradio
markupsafe==2.0.1
termcolor
jsonlines
mecab-python3
unidic-lite
einops
flash-attn
auto-gptq
fire
# alpaca-farm eval - my fork with some fixes.
git+https://github.com/hamishivi/alpaca_farm.git
# for human eval web app
flask
